Rename: api constant를 route constant에서 분리
