+++ webapp/src/constant/api.js
@@ -1,0 +1,3 @@
+export const ROOT_URL = process.env.REACT_APP_SERVER_API;
+
+export const MOCK_SERVER_URL = process.env.REACT_APP_MOCK_SERVER_API;
